(refactor) moved vitals fixture to accelerated directory
