+++ src/applications/mhv-medical-records/tests/e2e/accelerated/fixtures/vitals/body-weight.json
