+++ src/applications/mhv-medical-records/tests/e2e/accelerated/vitals/view-vitals-list.oracle-health.disabled.cypress.spec.js
@@ -1,7 +1,7 @@
 import MedicalRecordsSite from '../../mr_site/MedicalRecordsSite';
 import Vitals from '../pages/Vitals';
-import oracleHealthUser from '../../fixtures/user/oracle-health.json';
-import vitalsData from '../../fixtures/vitals/sample-lighthouse.json';
+import oracleHealthUser from '../fixtures/user/oracle-health.json';
+import vitalsData from '../fixtures/vitals/sample-lighthouse.json';
 
 describe('Medical Records View Vitals', () => {
   const site = new MedicalRecordsSite();
@@ -12,12 +12,15 @@ describe('Medical Records View Vitals', () => {
       isAcceleratingEnabled: false,
       isAcceleratingVitals: false,
     });
-    cy.visit('my-health/medical-records');
 
     Vitals.setIntercepts({ vitalData: vitalsData, useOhData: false });
   });
 
-  it('Visits View Vitals List', () => {
+  it('Visits View Vital List', () => {
+    site.loadPage();
+
+    cy.injectAxeThenAxeCheck();
+
     // check for MY Va Health links
     cy.get('[data-testid="labs-and-tests-landing-page-link"]').should(
       'be.visible',
@@ -28,12 +31,6 @@ describe('Medical Records View Vitals', () => {
 
     cy.get('[data-testid="allergies-landing-page-link"]').should('be.visible');
 
-    cy.get('[data-testid="vitals-landing-page-link"]')
-      .should('be.visible')
-      .click();
-
-    // Axe check
-    cy.injectAxe();
-    cy.axeCheck('main');
+    cy.get('[data-testid="vitals-landing-page-link"]').should('be.visible');
   });
 });
