+++ src/applications/mhv-medical-records/tests/e2e/accelerated/vitals/view-vitals-list.oracle-health.domain.body-height.cypress.spec.js
@@ -1,7 +1,7 @@
 import MedicalRecordsSite from '../../mr_site/MedicalRecordsSite';
 import Vitals from '../pages/Vitals';
-import oracleHealthUser from '../../fixtures/user/oracle-health.json';
-import vitalsData from '../../fixtures/vitals/body-height.json';
+import oracleHealthUser from '../fixtures/user/oracle-health.json';
+import vitalsData from '../fixtures/vitals/body-height.json';
 
 describe('Medical Records View Body Height', () => {
   const site = new MedicalRecordsSite();
@@ -15,15 +15,18 @@ describe('Medical Records View Body Height', () => {
     Vitals.setIntercepts({ vitalData: vitalsData });
   });
 
-  it('Visits View Vitals List', () => {
-    cy.visit('my-health/medical-records');
+  it('Visits View Vital List', () => {
+    site.loadPage();
 
-    Vitals.checkLandingPageLinks();
     // check for MY Va Health links
+    Vitals.checkLandingPageLinks();
+
     Vitals.goToVitalPage();
+
     // switch to march 2024
     Vitals.selectMonthAndYear({ month: '3', year: 2024 });
     Vitals.verifySelectedDate({ dateString: 'March 2024' });
+
     // check for latest id
     cy.get('[data-testid="vital-body-height-measurement"]').should(
       'be.visible',
