+++ src/applications/mhv-medical-records/tests/e2e/accelerated/vitals/view-vitals-list.oracle-health.domain.body-weight.cypress.spec.js
@@ -1,7 +1,7 @@
 import MedicalRecordsSite from '../../mr_site/MedicalRecordsSite';
 import Vitals from '../pages/Vitals';
-import oracleHealthUser from '../../fixtures/user/oracle-health.json';
-import vitalsData from '../../fixtures/vitals/body-weight.json';
+import oracleHealthUser from '../fixtures/user/oracle-health.json';
+import vitalsData from '../fixtures/vitals/body-weight.json';
 
 describe('Medical Records View Body Weight', () => {
   const site = new MedicalRecordsSite();
@@ -12,13 +12,15 @@ describe('Medical Records View Body Weight', () => {
       isAcceleratingEnabled: true,
       isAcceleratingVitals: true,
     });
-    cy.visit('my-health/medical-records');
     Vitals.setIntercepts({ vitalData: vitalsData });
   });
 
-  it('Visits View Vitals List', () => {
-    Vitals.checkLandingPageLinks();
+  it('Visits View Vital List', () => {
+    site.loadPage();
+
     // check for MY Va Health links
+    Vitals.checkLandingPageLinks();
+
     Vitals.goToVitalPage();
     // switch to march 2024
     Vitals.selectMonthAndYear({ month: '3', year: 2024 });
